Lamb's CAR test (1995).

A parametric test proposed by Lamb 1995 that examines whether or not the cumulative abnormal return (CAR) significantly differs from zero.

car_lamb(list_of_returns, car_start, car_end, percentage = 90)

Arguments

list_of_returns
a list of objects of S3 class returns, each element of which is treated as a security.

car_start
an object of Date class giving the first date of the CAR period.

car_end
an object of Date class giving the last date of the CAR period.

percentage
a lowest allowed percentage of non-missing observation for each day to be incorporated into CAR. The default value is 90 percent.

Value

A data frame of the following columns:

name: a name of the test, i.e. "car_lamb"

car_start: the first date of the CAR period

car_end: the last date of the CAR period

average_percentage: an average share of non-missing observations over the CAR period

car_mean: an average abnormal return over the CAR period

statistic: a test's statistic

number_of_days: the number of days in the CAR period

significance: a significance of the statistic

Details

This function performs a test proposed by Lamb 1995 to investigate whether CAR significantly differs from zero. This tests uses the variance, specified by Lamb 1995. The advantage of this test is allowance for correlated cross-sectional returns. The test statistic is close enough to statistic, produced by car_brown_warner_1985. The critical values are standard normal. The significance levels of \(\alpha\) are 0.1, 0.05, and 0.01 (marked respectively by *, **, and ***).

References

Lamb R.P. An Exposure-Based Analysis of Property-Liability Insurer Stock Values around Hurricane Andrew. Journal of Risk and Insurance, 62(1):111-123, 1995.

See also

Examples

if (FALSE) {
library("magrittr")
rates_indx <- get_prices_from_tickers("^GSPC",
                                      start = as.Date("2019-04-01"),
                                      end = as.Date("2020-04-01"),
                                      quote = "Close",
                                      retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous")
tickers <- c("AMZN", "ZM", "UBER", "NFLX", "SHOP", "FB", "UPWK")
get_prices_from_tickers(tickers,
                        start = as.Date("2019-04-01"),
                        end = as.Date("2020-04-01"),
                        quote = "Close",
                        retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous") %>%
    apply_market_model(regressor = rates_indx,
                       same_regressor_for_all = TRUE,
                       market_model = "sim",
                       estimation_method = "ols",
                       estimation_start = as.Date("2019-04-01"),
                       estimation_end = as.Date("2020-03-13")) %>%
    car_lamb(car_start = as.Date("2020-03-16"),
             car_end = as.Date("2020-03-20"))
}
## The result of the code above is equivalent to:
data(securities_returns)
car_lamb(
    list_of_returns = securities_returns,
    car_start = as.Date("2020-03-16"),
    car_end = as.Date("2020-03-20")
)

#>       name  car_start    car_end average_percentage   car_mean statistic
#> 1 car_lamb 2020-03-16 2020-03-20                100 0.02213319  3.846464
#>   number_of_days significance
#> 1              5          ***