Returns the result of given event study parametric tests.

Performs main parametric tests for each date in the event window and returns a data frame of their statistics and significance.

parametric_tests(list_of_returns, event_start, event_end, all = TRUE, tests)

Arguments

list_of_returns
a list of objects of S3 class returns, each element of which is treated as a security.

event_start
an object of Date class giving the first date of the event period.

event_end
an object of Date class giving the last date of the event period.

all
a logical vector of length one indicating whether all tests should be performed. The default value is TRUE.

tests
a list of tests' functions among brown_warner_1980, brown_warner_1985, t_test, patell, boehmer, and lamb.

Value

A data frame of the following columns:

date: a calendar date

weekday: a day of the week

percentage: a share of non-missing observations for a given day

mean: an average abnormal return

Various tests' statistics and significance

Details

parametric_tests performs given tests among brown_warner_1980, brown_warner_1985, t_test, patell, boehmer, lamb and merge result to a single data frame. If all = TRUE (the default value), the function ignores the value of tests.

References

Brown S.J., Warner J.B. Measuring security price performance. Journal of Financial Economics, 8:205-258, 1980.

Brown S.J., Warner J.B. Using Daily Stock Returns, The Case of Event Studies. Journal of Financial Economics, 14:3-31, 1985.

Boehmer E., Musumeci J., Poulsen A.B. Event-study methodology under conditions of event-induced variance. Journal of Financial Economics, 30(2):253-272, 1991.

Patell J.M. Corporate forecasts of earnings per share and stock price behavior: empirical tests. Journal of Accounting Research, 14(2):246- 276, 1976.

Lamb R.P. An Exposure-Based Analysis of Property-Liability Insurer Stock Values around Hurricane Andrew. Journal of Risk and Insurance, 62(1):111-123, 1995.

See also

Examples

if (FALSE) {
library("magrittr")
rates_indx <- get_prices_from_tickers("^GSPC",
                                      start = as.Date("2019-04-01"),
                                      end = as.Date("2020-04-01"),
                                      quote = "Close",
                                      retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous")
tickers <- c("AMZN", "ZM", "UBER", "NFLX", "SHOP", "FB", "UPWK")
param <- get_prices_from_tickers(tickers,
                                 start = as.Date("2019-04-01"),
                                 end = as.Date("2020-04-01"),
                                 quote = "Close",
                                 retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous") %>%
    apply_market_model(regressor = rates_indx,
                       same_regressor_for_all = TRUE,
                       market_model = "sim",
                       estimation_method = "ols",
                       estimation_start = as.Date("2019-04-01"),
                       estimation_end = as.Date("2020-03-13")) %>%
    parametric_tests(event_start = as.Date("2020-03-16"),
                     event_end = as.Date("2020-03-20"))
}
## The result of the code above is equivalent to:
data(securities_returns)
param <- parametric_tests(list_of_returns = securities_returns,
                          event_start = as.Date("2020-03-16"),
                          event_end = as.Date("2020-03-20"))