Patell's parametric test (1976).

An event study parametric test described in Patell 1976.

patell(list_of_returns, event_start, event_end)

Arguments

list_of_returns
a list of objects of S3 class returns, each element of which is treated as a security.

event_start
an object of Date class giving the first date of the event period.

event_end
an object of Date class giving the last date of the event period.

Value

A data frame of the following columns:

date: a calendar date

weekday: a day of the week

percentage: a share of non-missing observations for a given day

mean: an average abnormal return

pt_stat: a Patell's test statistic

pt_signif: a significance of the statistic

Details

Performs a parametric test for event study, which is described in Patell 1976, which is called standardized-residuals method in Boehmer 1991. Test's assumptions are a cross-sectional independence and an insignificance of an event-induced variance. The standardization smooths the effect of the event-induced variance comparing to Brown and Warner tests. Also standardization incorporates the situation, when a highly volatile security dominates the test. The test examines the hypothesis whether the theoretical cross-sectional expected value for a given day is equal to zero. It calculates statistics even if event window and estimation period are overlapped (intersect). The critical values are standard normal. The significance levels of \(\alpha\) are 0.1, 0.05, and 0.01 (marked respectively by *, **, and ***).

References

Patell J.M. Corporate forecasts of earnings per share and stock price behavior: empirical tests. Journal of Accounting Research, 14(2):246- 276, 1976.

Boehmer E., Musumeci J., Poulsen A.B. Event-study methodology under conditions of event-induced variance. Journal of Financial Economics, 30(2):253-272, 1991.

See also

Examples

if (FALSE) {
library("magrittr")
rates_indx <- get_prices_from_tickers("^N100",
                                      start = as.Date("2000-01-01"),
                                      end = as.Date("2002-01-01"),
                                      quote = "Close",
                                      retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous")
tickers <- c("ALV.DE", "CS.PA", "G.MI", "HNR1.HA", "HSX.L", "MUV2.DE",
             "TOP.CO")
get_prices_from_tickers(tickers,
                        start = as.Date("2000-01-01"),
                        end = as.Date("2002-01-01"),
                        quote = "Close",
                        retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous") %>%
    apply_market_model(regressor = rates_indx,
                       same_regressor_for_all = TRUE,
                       market_model = "sim",
                       estimation_method = "ols",
                       estimation_start = as.Date("2001-03-26"),
                       estimation_end = as.Date("2001-09-10")) %>%
    patell(event_start = as.Date("2001-09-11"),
           event_end = as.Date("2001-09-28"))
}
## The result of the code above is equivalent to:
data(securities_returns)
patell(list_of_returns = securities_returns,
       event_start =  as.Date("2001-09-11"),
       event_end = as.Date("2001-09-28"))

#>          date   weekday percentage         mean      pt_stat pt_signif
#> 1  2001-09-11   Tuesday        100 -0.095267547 -15.93199210       ***
#> 2  2001-09-12 Wednesday        100 -0.031384102  -3.02501512       ***
#> 3  2001-09-13  Thursday          0           NA           NA          
#> 4  2001-09-14    Friday        100 -0.037994921  -6.04629140       ***
#> 5  2001-09-17    Monday        100  0.023717979   4.87549214       ***
#> 6  2001-09-18   Tuesday        100 -0.004574035  -0.72885915          
#> 7  2001-09-19 Wednesday        100 -0.004380471  -0.01540181          
#> 8  2001-09-20  Thursday        100 -0.097628194 -13.28950663       ***
#> 9  2001-09-21    Friday        100 -0.039622894  -6.18660223       ***
#> 10 2001-09-24    Monday        100  0.031556552   4.66791310       ***
#> 11 2001-09-25   Tuesday        100  0.041640460   6.22810566       ***
#> 12 2001-09-26 Wednesday        100  0.022896229   3.48075345       ***
#> 13 2001-09-27  Thursday        100  0.020423192   3.32642949       ***
#> 14 2001-09-28    Friday        100  0.005078339   1.02097492