An event study simple binomial sign test.

A binomial sign test which determines whether the frequency of positive abnormal returns in the event period is significantly different from one-half.

sign_test(list_of_returns, event_start, event_end)

Arguments

list_of_returns
a list of objects of S3 class returns, each element of which is treated as a security.

event_start
an object of Date class giving the first date of the event period.

event_end
an object of Date class giving the last date of the event period.

Value

A data frame of the following columns:

date: a calendar date

weekday: a day of the week

percentage: a share of non-missing observations for a given day

sign_stat: a sign test statistic

sign_signif: a significance of the statistic

Details

This test is application of the simple binomial test to the event study, which indicates whether the cross-sectional frequency of positive abnormal returns is significantly different from 0.5. This test is stable to outliers, in other words allows for checking if the result is driven by few companies with extremely large abnormal performance. For this test the estimation period and the event period must not overlap, otherwise an error will be thrown. The test statistic is assumed to have a normal distribution in approximation under a null hypothesis, if the number of securities is large. Typically the test is used together with parametric tests. The test is well-specified for the case, when cross-sectional abnormal returns are not symmetric. Also this procedure is less sensitive to extreme returns than the rank test. The significance levels of \(\alpha\) are 0.1, 0.05, and 0.01 (marked respectively by *, **, and ***).

References

Boehmer E., Musumeci J., Poulsen A.B. Event-study methodology under conditions of event-induced variance. Journal of Financial Economics, 30(2):253-272, 1991.

See also

Examples

if (FALSE) {
library("magrittr")
rates_indx <- get_prices_from_tickers("^N100",
                                      start = as.Date("2000-01-01"),
                                      end = as.Date("2002-01-01"),
                                      quote = "Close",
                                      retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous")
tickers <- c("ALV.DE", "CS.PA", "G.MI", "HNR1.HA", "HSX.L", "MUV2.DE",
             "RSA.L", "TOP.CO")
get_prices_from_tickers(tickers,
                        start = as.Date("2000-01-01"),
                        end = as.Date("2002-01-01"),
                        quote = "Close",
                        retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous") %>%
    apply_market_model(regressor = rates_indx,
                       same_regressor_for_all = TRUE,
                       market_model = "sim",
                       estimation_method = "ols",
                       estimation_start = as.Date("2001-03-26"),
                       estimation_end = as.Date("2001-09-10")) %>%
    sign_test(event_start = as.Date("2001-09-11"),
              event_end = as.Date("2001-09-28"))
}
## The result of the code above is equivalent to:
data(securities_returns)
sign_test(list_of_returns = securities_returns,
          event_start =  as.Date("2001-09-11"),
          event_end = as.Date("2001-09-28"))

#>          date   weekday percentage  sign_stat sign_signif
#> 1  2001-09-11   Tuesday        100 -2.8284271         ***
#> 2  2001-09-12 Wednesday        100 -1.4142136            
#> 3  2001-09-13  Thursday          0         NA            
#> 4  2001-09-14    Friday        100 -1.4142136            
#> 5  2001-09-17    Monday        100  0.7071068            
#> 6  2001-09-18   Tuesday        100  0.7071068            
#> 7  2001-09-19 Wednesday        100 -0.7071068            
#> 8  2001-09-20  Thursday        100 -2.8284271         ***
#> 9  2001-09-21    Friday        100 -1.4142136            
#> 10 2001-09-24    Monday        100  2.1213203          **
#> 11 2001-09-25   Tuesday        100  0.7071068            
#> 12 2001-09-26 Wednesday        100  2.1213203          **
#> 13 2001-09-27  Thursday        100  0.7071068            
#> 14 2001-09-28    Friday         75  1.9051587           *