An event study Wilcoxon signed rank test.

Performs Wilcoxon test on the event period for abnormal returns (abnormal returns are considered as differences).

wilcoxon_test(list_of_returns, event_start, event_end)

Arguments

list_of_returns
a list of objects of S3 class returns, each element of which is treated as a security.

event_start
an object of Date class giving the first date of the event period.

event_end
an object of Date class giving the last date of the event period.

Value

A data frame of the following columns:

date: a calendar date

weekday: a day of the week

percentage: a share of non-missing observations for a given day

wlcx_stat: a Wilcoxon signed rank test statistic

wlcx_signif: a significance of the statistic

Details

The estimation periods can overlap with event windows, because the procedure takes into account only abnormal returns from the event window. The test has the same algorithm as built-in R wilcox.test. The critical values are exact values, which are obtained from qsignrank. The algorithm is the following: for each day in event window the cross-sectional abnormal returns treated as sample of differences. Firstly the absolute value of these differences are computed, and corresponding ranks of non-zero values are calculated. The test statistic is the sum of ranks, corresponding to positive abnormal returns. The significance levels of \(\alpha\) are 0.1, 0.05, and 0.01 (marked respectively by *, **, and ***).

References

Wilcoxon F. Individual Comparisons by Ranking Methods. Biometrics Bulletin 1(6):80-83, 1945.

Kolari J.W., Pynnonen S. Event Study Testing with Cross-sectional Correlation of Abnormal Returns. The Review of Financial Studies, 23(11):3996-4025, 2010.

Lehmann E.L, Nonparametrics: Statistical Methods Based on Ranks. San Francisco: Holden-Day, 1975.

Hollander M., Wolfe D.A. Nonparametric Statistical Methods. New York: John Wiley & Sons, 1973.

See also

Examples

if (FALSE) {
library("magrittr")
rates_indx <- get_prices_from_tickers("^STOXX50E",
                                      start = as.Date("2000-01-01"),
                                      end = as.Date("2002-01-01"),
                                      quote = "Close",
                                      retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous")
tickers <- c("ALV.DE", "CS.PA", "G.MI", "HNR1.HA", "HSX.L", "MUV2.DE",
             "RSA.L", "TOP.CO")
get_prices_from_tickers(tickers,
                        start = as.Date("2000-01-01"),
                        end = as.Date("2002-01-01"),
                        quote = "Close",
                        retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous") %>%
    apply_market_model(regressor = rates_indx,
                       same_regressor_for_all = TRUE,
                       market_model = "sim",
                       estimation_method = "ols",
                       estimation_start = as.Date("2001-03-26"),
                       estimation_end = as.Date("2001-09-10")) %>%
    wilcoxon_test(event_start = as.Date("2001-09-11"),
                  event_end = as.Date("2001-09-28"))
}
## The result of the code above is equivalent to:
data(securities_returns)
wilcoxon_test(list_of_returns = securities_returns,
              event_start =  as.Date("2001-09-11"),
              event_end = as.Date("2001-09-28"))

#>          date   weekday percentage wlcx_stat wlcx_signif
#> 1  2001-09-11   Tuesday        100         0         ***
#> 2  2001-09-12 Wednesday        100         4          **
#> 3  2001-09-13  Thursday        100        19            
#> 4  2001-09-14    Friday        100        10            
#> 5  2001-09-17    Monday        100        16            
#> 6  2001-09-18   Tuesday        100        16            
#> 7  2001-09-19 Wednesday        100        15            
#> 8  2001-09-20  Thursday        100         0         ***
#> 9  2001-09-21    Friday        100         3          **
#> 10 2001-09-24    Monday        100        30          **
#> 11 2001-09-25   Tuesday        100        31          **
#> 12 2001-09-26 Wednesday        100        33          **
#> 13 2001-09-27  Thursday        100        22            
#> 14 2001-09-28    Friday         75        24         ***