Boehmer's parametric test (1991).

An event study parametric test described in Boehmer 1991.

boehmer(list_of_returns, event_start, event_end)

Arguments

list_of_returns
a list of objects of S3 class returns, each element of which is treated as a security.

event_start
an object of Date class giving the first date of the event period.

event_end
an object of Date class giving the last date of the event period.

Value

A data frame of the following columns:

date: a calendar date

weekday: a day of the week

percentage: a share of non-missing observations for a given day

mean: an average abnormal return

bh_stat: a Boehmer's test statistic

bh_signif: a significance of the statistic

Details

Performs a parametric test for event study, which is described in Boehmer 1991. Also called hybrid test or standardized cross-sectional test. This test performs t-test based on Patell's standardized residuals. By combining Patell's and t-tests, this test allows for event-induced variance changes, but still assumes cross-sectional independence. The test examines the hypothesis whether the theoretical cross-sectional expected value for a given day is equal to zero. It calculates statistics even if event window and estimation period are overlapped (intersect). The critical values has Student's t-distribution. The significance levels of \(\alpha\) are 0.1, 0.05, and 0.01 (marked respectively by *, **, and ***).

References

Patell J.M. Corporate forecasts of earnings per share and stock price behavior: empirical tests. Journal of Accounting Research, 14(2):246- 276, 1976.

Boehmer E., Musumeci J., Poulsen A.B. Event-study methodology under conditions of event-induced variance. Journal of Financial Economics, 30(2):253-272, 1991.

See also

Examples

if (FALSE) {
library("magrittr")
rates_indx <- get_prices_from_tickers("^N100",
                                      start = as.Date("2000-01-01"),
                                      end = as.Date("2002-01-01"),
                                      quote = "Close",
                                      retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous")
tickers <- c("ALV.DE", "CS.PA", "G.MI", "HNR1.HA", "HSX.L", "MUV2.DE",
             "RSA.L", "TOP.CO")
get_prices_from_tickers(tickers,
                        start = as.Date("2000-01-01"),
                        end = as.Date("2002-01-01"),
                        quote = "Close",
                        retclass = "zoo") %>%
    get_rates_from_prices(quote = "Close",
                          multi_day = TRUE,
                          compounding = "continuous") %>%
    apply_market_model(regressor = rates_indx,
                       same_regressor_for_all = TRUE,
                       market_model = "sim",
                       estimation_method = "ols",
                       estimation_start = as.Date("2001-03-26"),
                       estimation_end = as.Date("2001-09-10")) %>%
    boehmer(event_start = as.Date("2001-09-11"),
            event_end = as.Date("2001-09-28"))
}
## The result of the code above is equivalent to:
data(securities_returns)
boehmer(list_of_returns = securities_returns,
        event_start =  as.Date("2001-09-11"),
        event_end = as.Date("2001-09-28"))

#>          date   weekday percentage          mean     bh_stat bh_signif
#> 1  2001-09-11   Tuesday        100 -0.1034808908 -4.45409291       ***
#> 2  2001-09-12 Wednesday        100 -0.0345902954 -0.98762098          
#> 3  2001-09-13  Thursday          0            NA          NA          
#> 4  2001-09-14    Friday        100 -0.0340256202 -1.42803620          
#> 5  2001-09-17    Monday        100  0.0180458400  1.14070051          
#> 6  2001-09-18   Tuesday        100 -0.0007609145 -0.03168897          
#> 7  2001-09-19 Wednesday        100 -0.0048350600 -0.02820835          
#> 8  2001-09-20  Thursday        100 -0.0974285486 -3.03467540        **
#> 9  2001-09-21    Friday        100 -0.0412253278 -2.75369055        **
#> 10 2001-09-24    Monday        100  0.0369475589  1.91033095         *
#> 11 2001-09-25   Tuesday        100  0.0467385257  2.61279384        **
#> 12 2001-09-26 Wednesday        100  0.0188989270  3.49544565        **
#> 13 2001-09-27  Thursday        100  0.0112991782  1.06757580          
#> 14 2001-09-28    Friday         75  0.0124532411  2.33716388         *